DoEffect reusable by Au4
