+++ libraries/lib-audacity-application-logic/CompressorInstance.h
@@ -11,6 +11,7 @@
 #pragma once
 
 #include "DynamicRangeProcessorTypes.h"
+#include "CompressorProcessor.h"
 #include "EffectInterface.h"
 #include "Observer.h"
 #include "PerTrackEffect.h"
@@ -18,13 +19,11 @@
 #include <optional>
 #include <vector>
 
-class CompressorProcessor;
-
 using InitializeProcessingSettingsPublisher =
    Observer::Publisher<std::optional<InitializeProcessingSettings>>;
 using RealtimeResumePublisher = Observer::Publisher<Unbypassed>;
 
-class CompressorInstance final :
+class AUDACITY_APPLICATION_LOGIC_API CompressorInstance final :
     public PerTrackEffect::Instance,
     public EffectInstanceWithBlockSize,
     public InitializeProcessingSettingsPublisher,
