+++ libraries/lib-components/EffectInterface.h
@@ -51,7 +51,6 @@
 #include <memory>
 #include <optional>
 #include <type_traits>
-#include <wx/event.h>
 
 #include "Observer.h"
 
@@ -405,9 +404,6 @@ public:
    virtual std::unique_ptr<EffectOutputs> MakeOutputs() const;
    //! @}
 };
-
-class wxDialog;
-class wxWindow;
 
 // ----------------------------------------------------------------------------
 // Supported channel assignments
