+++ src/BatchProcessDialog.cpp
@@ -14,6 +14,7 @@
 
 *//*******************************************************************/
 #include "BatchProcessDialog.h"
+#include "DoEffect.h"
 
 #include <wx/setup.h> // for wxUSE_* macros
 
@@ -35,6 +36,7 @@
 #include <wx/settings.h>
 
 #include "Clipboard.h"
+#include "PluginManager.h"
 #include "ShuttleGui.h"
 #include "MenuCreator.h"
 #include "Prefs.h"
@@ -45,9 +47,9 @@
 #include "ProjectWindows.h"
 #include "SelectUtilities.h"
 #include "Track.h"
+
 #include "CommandManager.h"
 #include "Effect.h"
-#include "effects/EffectManager.h"
 #include "effects/EffectUI.h"
 #include "../images/Arrow.xpm"
 #include "../images/Empty9x16.xpm"
@@ -895,8 +897,8 @@ void MacrosWindow::OnListSelected(wxListEvent &event)
 
    if (command != mCatalog.end())
    {
-      EffectManager &em = EffectManager::Get();
-      PluginID ID = em.GetEffectByIdentifier(command->name.Internal());
+      PluginID ID =
+         PluginManager::Get().GetByCommandIdentifier(command->name.Internal());
 
       mEdit->Enable(!ID.empty());
    }
@@ -1231,7 +1233,7 @@ void MacrosWindow::OnEditCommandParams(wxCommandEvent & WXUNUSED(event))
    wxString params  = mMacroCommands.GetParams(item);
    wxString oldParams = params;
 
-   params = MacroCommands::PromptForParamsFor(command, params, *this).Trim();
+   params = MacroCommands::PromptForParamsFor(command, params, mProject).Trim();
    Raise();
 
    if (oldParams == params)
@@ -1396,7 +1398,6 @@ void MacrosWindow::UpdatePrefs()
 
 #include "CommonCommandFlags.h"
 #include "CommandContext.h"
-#include "effects/EffectManager.h"
 namespace {
 
 AttachedWindows::RegisteredFactory sMacrosWindowKey{
@@ -1421,10 +1422,8 @@ void OnRepeatLastTool(const CommandContext& context)
      {
         auto lastEffect = commandManager.mLastTool;
         if (!lastEffect.empty())
-        {
            EffectUI::DoEffect(
-              lastEffect, context, commandManager.mRepeatToolFlags);
-        }
+              lastEffect, context.project, commandManager.mRepeatToolFlags);
      }
        break;
      case CommandManager::repeattypeunique:
@@ -1493,8 +1492,7 @@ void OnApplyMacroDirectlyByName(const CommandContext& context, const MacroID& Na
    CommandManager::Get(project).ModifyUndoMenuItems();
 
    TranslatableString desc;
-   EffectManager& em = EffectManager::Get();
-   auto shortDesc = em.GetCommandName(Name);
+   auto shortDesc = PluginManager::Get().GetName(Name);
    auto& undoManager = UndoManager::Get(project);
    auto& commandManager = CommandManager::Get(project);
    int cur = undoManager.GetCurrentState();
@@ -1551,8 +1549,9 @@ auto PluginMenuItems()
             const auto &lastTool = CommandManager::Get(project).mLastTool;
             TranslatableString buildMenuLabel;
             if (!lastTool.empty())
-               buildMenuLabel = XO("Repeat %s")
-                  .Format( EffectManager::Get().GetCommandName(lastTool) );
+               buildMenuLabel =
+                  XO("Repeat %s")
+                     .Format(PluginManager::Get().GetName(lastTool));
             else
                buildMenuLabel = XO("Repeat Last Tool");
