+++ src/CMakeLists.txt
@@ -105,6 +105,8 @@ set( SOURCES
       Diags.cpp
       Diags.h
       DropTarget.cpp
+      EffectAndCommandPluginManager.cpp
+      EffectAndCommandPluginManager.h
       DropoutDetector.cpp
       EnvelopeEditor.cpp
       EnvelopeEditor.h
@@ -134,6 +136,8 @@ set( SOURCES
       LangChoice.h
       Legacy.cpp
       Legacy.h
+      DoEffect.cpp
+      DoEffect.h
       ListNavigationEnabled.cpp
       ListNavigationEnabled.h
       ListNavigationPanel.cpp
@@ -360,10 +364,10 @@ set( SOURCES
       effects/Echo.h
       effects/EffectEditor.cpp
       effects/EffectEditor.h
-      effects/EffectManager.cpp
-      effects/EffectManager.h
       effects/EffectPreview.cpp
       effects/EffectPreview.h
+      effects/EffectPresetDialog.cpp
+      effects/EffectPresetDialog.h
       effects/EffectUI.cpp
       $<$<BOOL:${wxIS_MAC}>:effects/EffectUI.mm>
       effects/EffectUI.h
@@ -417,8 +421,6 @@ set( SOURCES
       effects/Compressor.h
       effects/CompressorEditor.cpp
       effects/CompressorEditor.h
-      effects/CompressorInstance.cpp
-      effects/CompressorInstance.h
       effects/RealtimeEffectStateUI.cpp
       effects/RealtimeEffectStateUI.h
       effects/Repair.cpp
@@ -1323,21 +1325,16 @@ set( AUDACITY_LIBRARIES
    lib-theme-resources-interface
    lib-graphics-interface
    lib-tags-interface
-   lib-project-file-io-interface
-   lib-effects-interface
    lib-wx-init-interface
    lib-time-track-interface
-   lib-audio-io-interface
    lib-stretching-sequence-interface
    lib-import-export-interface
    lib-export-ui-interface
-   lib-menus-interface
    lib-note-track-interface
-   lib-viewport-interface
    lib-wave-track-paint-interface
    lib-music-information-retrieval-interface
    lib-preference-pages-interface
-   lib-dynamic-range-processor-interface
+   lib-audacity-application-logic-interface
 )
 
 if (USE_VST)
