+++ src/commands/HelpCommand.cpp
@@ -17,15 +17,16 @@
 
 #include "HelpCommand.h"
 
-#include "CommandDispatch.h"
-#include "MenuRegistry.h"
 #include "../CommonCommandFlags.h"
-#include "SettingsVisitor.h"
-#include "LoadCommands.h"
-#include "ShuttleGui.h"
-#include "CommandTargets.h"
 #include "CommandContext.h"
-#include "../effects/EffectManager.h"
+#include "CommandDispatch.h"
+#include "CommandTargets.h"
+#include "EffectAndCommandPluginManager.h"
+#include "LoadCommands.h"
+#include "MenuRegistry.h"
+#include "PluginManager.h"
+#include "SettingsVisitor.h"
+#include "ShuttleGui.h"
 
 const ComponentInterfaceSymbol HelpCommand::Symbol
 { XO("Help") };
@@ -46,7 +47,7 @@ enum {
 static const EnumValueSymbol kFormats[nFormats] =
 {
    // These are acceptable dual purpose internal/visible names
-   
+
    /* i18n-hint JavaScript Object Notation */
    { XO("JSON") },
    /* i18n-hint name of a computer programming language */
@@ -86,7 +87,7 @@ bool HelpCommand::Apply(const CommandContext &context)
 
    if( mFormat == kLisp )
    {
-      CommandContext LispyContext( 
+      CommandContext LispyContext(
          context.project,
          std::make_unique<LispifiedCommandOutputTargets>( *context.pOutput.get() )
          );
@@ -95,7 +96,7 @@ bool HelpCommand::Apply(const CommandContext &context)
 
    if( mFormat == kBrief )
    {
-      CommandContext BriefContext( 
+      CommandContext BriefContext(
          context.project,
          std::make_unique<BriefCommandOutputTargets>( *context.pOutput.get() )
          );
@@ -106,12 +107,11 @@ bool HelpCommand::Apply(const CommandContext &context)
 }
 
 bool HelpCommand::ApplyInner(const CommandContext & context){
-   EffectManager & em = EffectManager::Get();
-   PluginID ID = em.GetEffectByIdentifier( mCommandName );
+   PluginID ID = PluginManager::Get().GetByCommandIdentifier(mCommandName);
    if( ID.empty() )
       context.Status( "Command not found" );
    else
-      em.GetCommandDefinition( ID, context, 1);
+      EffectAndCommandPluginManager::Get().GetCommandDefinition(ID, context, 1);
    return true;
 }
