+++ src/effects/EffectUI.cpp
@@ -22,6 +22,7 @@
 #include "CommandManager.h"
 #include "ConfigInterface.h"
 #include "EffectManager.h"
+#include "DoEffect.h"
 #include "PluginManager.h"
 #include "ProjectAudioIO.h"
 #include "ProjectHistory.h"
@@ -104,7 +105,6 @@ private:
 #include "../ProjectAudioManager.h"
 #include "ShuttleGui.h"
 #include "ViewInfo.h"
-#include "../commands/AudacityCommand.h"
 #include "CommandContext.h"
 #include "AudacityMessageBox.h"
 #include "HelpSystem.h"
@@ -252,7 +252,7 @@ EffectUIHost::EffectUIHost(wxWindow *parent,
 #if defined(__WXMAC__)
    MacMakeWindowFloating(GetHandle());
 #endif
-   
+
    SetName( effect.GetDefinition().GetName() );
 
    // This style causes Validate() and TransferDataFromWindow() to visit
@@ -297,7 +297,7 @@ bool EffectUIHost::TransferDataFromWindow()
    //! Do other custom validation and transfer actions
    if (!mpEditor->ValidateUI())
       return false;
-   
+
    // Transfer-from takes non-const reference to settings
    bool result = true;
    mpAccess->ModifySettings([&](EffectSettings &settings){
@@ -378,7 +378,7 @@ void EffectUIHost::BuildTopBar(ShuttleGui &S)
 
          S.AddSpace(1, 0, 1);
 
-         
+
 
          if (mEffectUIHost.GetDefinition().EnablesDebug())
          {
@@ -451,7 +451,7 @@ bool EffectUIHost::Initialize()
                S.AddSpace(1, 1, 1);
                S.Id(wxID_CANCEL)
                   .AddButton(XXO("&Cancel"));
-               
+
                mApplyBtn = S.Id(wxID_APPLY)
                   .AddButton( XXO("&Apply"));
                mApplyBtn->SetDefault();
@@ -495,11 +495,11 @@ void EffectUIHost::OnInitDialog(wxInitDialogEvent & evt)
 {
    // Do default handling
    wxDialogWrapper::OnInitDialog(evt);
-   
+
 #if wxCHECK_VERSION(3, 0, 0)
    //#warning "check to see if this still needed in wx3"
 #endif
-   
+
    // Pure hackage coming down the pike...
    //
    // I have no idea why, but if a wxTextCtrl is the first control in the
@@ -522,7 +522,7 @@ void EffectUIHost::OnErase(wxEraseEvent & WXUNUSED(evt))
 void EffectUIHost::OnPaint(wxPaintEvent & WXUNUSED(evt))
 {
    wxPaintDC dc(this);
-   
+
    dc.Clear();
 }
 
@@ -533,10 +533,10 @@ void EffectUIHost::OnClose(wxCloseEvent & WXUNUSED(evt))
 
    if (mpEditor)
       mpEditor->OnClose();
-   
+
    Hide();
    Destroy();
-   
+
 #if wxDEBUG_LEVEL
    mClosed = true;
 #endif
@@ -571,7 +571,7 @@ void EffectUIHost::OnApply(wxCommandEvent & evt)
       if (!allowed)
          return;
    }
-   
+
    if (!TransferDataFromWindow() ||
        // This is the main place where there is a side-effect on the config
        // file to remember the last-used settings of an effect, just before
@@ -583,14 +583,14 @@ void EffectUIHost::OnApply(wxCommandEvent & evt)
    if (IsModal())
    {
       mDismissed = true;
-      
+
       EndModal(evt.GetId());
-      
+
       Close();
-      
+
       return;
    }
-   
+
    // Progress dialog no longer yields, so this "shouldn't" be necessary (yet to be proven
    // for sure), but it is a nice visual cue that something is going on.
    mApplyBtn->Disable();
@@ -600,7 +600,7 @@ void EffectUIHost::OnApply(wxCommandEvent & evt)
    // This is absolute hackage...but easy and I can't think of another way just now.
    //
    // It should callback to the EffectManager to kick off the processing
-   EffectUI::DoEffect(GetID(mEffectUIHost), context,
+   EffectUI::DoEffect(GetID(mEffectUIHost), context.project,
       EffectManager::kConfigured);
 }
 
@@ -620,7 +620,7 @@ void EffectUIHost::DoCancel()
          EndModal(0);
       else
          Hide();
-      
+
       mDismissed = true;
    }
 }
@@ -653,7 +653,7 @@ void EffectUIHost::OnMenu(wxCommandEvent & WXUNUSED(evt))
    menu.Bind(wxEVT_MENU, [](auto&){}, kUserPresetsDummyID);
    menu.Bind(wxEVT_MENU, [](auto&){}, kDeletePresetDummyID);
    LoadUserPresets();
-   
+
    if (mUserPresets.size() == 0)
    {
       menu.Append(kUserPresetsDummyID, _("User Presets"))->Enable(false);
@@ -667,9 +667,9 @@ void EffectUIHost::OnMenu(wxCommandEvent & WXUNUSED(evt))
       }
       menu.Append(0, _("User Presets"), sub.release());
    }
-   
+
    menu.Append(kSaveAsID, _("Save Preset..."));
-   
+
    if (mUserPresets.size() == 0)
    {
       menu.Append(kDeletePresetDummyID, _("Delete Preset"))->Enable(false);
@@ -683,11 +683,11 @@ void EffectUIHost::OnMenu(wxCommandEvent & WXUNUSED(evt))
       }
       menu.Append(0, _("Delete Preset"), sub.release());
    }
-   
+
    menu.AppendSeparator();
-   
+
    auto factory = mEffectUIHost.GetDefinition().GetFactoryPresets();
-   
+
    {
       auto sub = std::make_unique<wxMenu>();
       sub->Append(kDefaultsID, _("Defaults"));
@@ -701,13 +701,13 @@ void EffectUIHost::OnMenu(wxCommandEvent & WXUNUSED(evt))
             {
                label = _("None");
             }
-            
+
             sub->Append(kFactoryPresetsID + i, label);
          }
       }
       menu.Append(0, _("Factory Presets"), sub.release());
    }
-   
+
    menu.AppendSeparator();
    menu.Append(kImportID, _("Import..."))
       ->Enable(mEffectUIHost.CanExportPresets());
@@ -717,10 +717,10 @@ void EffectUIHost::OnMenu(wxCommandEvent & WXUNUSED(evt))
    menu.Append(kOptionsID, _("Options..."))
       ->Enable(mEffectUIHost.HasOptions());
    menu.AppendSeparator();
-   
+
    {
       auto sub = std::make_unique<wxMenu>();
-      
+
       auto &definition = mEffectUIHost.GetDefinition();
       sub->Append(kDummyID, wxString::Format(_("Type: %s"),
          ::wxGetTranslation( definition.GetFamily().Translation() )));
@@ -733,7 +733,7 @@ void EffectUIHost::OnMenu(wxCommandEvent & WXUNUSED(evt))
 
       menu.Append(0, _("About"), sub.release());
    }
-   
+
    wxWindow *btn = FindWindow(kMenuID);
    wxRect r = btn->GetRect();
    BasicMenu::Handle{ &menu }.Popup(
@@ -759,7 +759,7 @@ void EffectUIHost::OnPlay(wxCommandEvent & WXUNUSED(evt))
 {
    if (!TransferDataFromWindow())
       return;
-   
+
    auto updater = [this]{ TransferDataToWindow(); };
    EffectPreview(mEffectUIHost, *mpAccess, updater, false);
    // After restoration of settings and effect state:
@@ -784,7 +784,7 @@ void EffectUIHost::OnCapture(AudioIOEvent evt)
 void EffectUIHost::OnUserPreset(wxCommandEvent & evt)
 {
    int preset = evt.GetId() - kUserPresetsID;
-   
+
    mpAccess->ModifySettings([&](EffectSettings &settings){
       // ignore failure
       return mEffectUIHost.GetDefinition().LoadUserPreset(
@@ -808,7 +808,7 @@ void EffectUIHost::OnFactoryPreset(wxCommandEvent & evt)
 void EffectUIHost::OnDeletePreset(wxCommandEvent & evt)
 {
    auto preset = mUserPresets[evt.GetId() - kDeletePresetID];
-   
+
    int res = AudacityMessageBox(
                                 XO("Are you sure you want to delete \"%s\"?").Format( preset ),
                                 XO("Delete Preset"),
@@ -818,9 +818,9 @@ void EffectUIHost::OnDeletePreset(wxCommandEvent & evt)
       RemoveConfigSubgroup(mEffectUIHost.GetDefinition(),
          PluginSettings::Private, UserPresetsGroup(preset));
    }
-   
+
    LoadUserPresets();
-   
+
    return;
 }
 
@@ -829,9 +829,9 @@ void EffectUIHost::OnSaveAs(wxCommandEvent & WXUNUSED(evt))
    wxTextCtrl *text;
    wxString name;
    wxDialogWrapper dlg(this, wxID_ANY, XO("Save Preset"));
-   
+
    ShuttleGui S(&dlg, eIsCreating);
-   
+
    S.StartPanel();
    {
       S.StartVerticalLay(1);
@@ -847,20 +847,20 @@ void EffectUIHost::OnSaveAs(wxCommandEvent & WXUNUSED(evt))
       S.EndVerticalLay();
    }
    S.EndPanel();
-   
+
    dlg.SetSize(dlg.GetSizer()->GetMinSize());
    dlg.Center();
    dlg.Fit();
-   
+
    while (true)
    {
       int rc = dlg.ShowModal();
-      
+
       if (rc != wxID_OK)
       {
          break;
       }
-      
+
       name = text->GetValue();
       if (name.empty())
       {
@@ -872,7 +872,7 @@ void EffectUIHost::OnSaveAs(wxCommandEvent & WXUNUSED(evt))
          md.ShowModal();
          continue;
       }
-      
+
       if ( make_iterator_range( mUserPresets ).contains( name ) )
       {
          AudacityMessageDialog md(
@@ -886,21 +886,21 @@ void EffectUIHost::OnSaveAs(wxCommandEvent & WXUNUSED(evt))
          {
             break;
          }
-         
+
          if (choice == wxID_NO)
          {
             continue;
          }
       }
-      
+
       if (TransferDataFromWindow())
          mEffectUIHost.GetDefinition()
             .SaveUserPreset(UserPresetsGroup(name), mpAccess->Get());
       LoadUserPresets();
-      
+
       break;
    }
-   
+
    return;
 }
 
@@ -922,14 +922,14 @@ void EffectUIHost::OnExport(wxCommandEvent & WXUNUSED(evt))
    // exceptions are handled in AudacityApp::OnExceptionInMainLoop
    if (TransferDataFromWindow())
      mClient.ExportPresets(mEffectUIHost, mpAccess->Get());
-   
+
    return;
 }
 
 void EffectUIHost::OnOptions(wxCommandEvent & WXUNUSED(evt))
 {
    mClient.ShowOptions(mEffectUIHost);
-   
+
    return;
 }
 
@@ -958,7 +958,7 @@ void EffectUIHost::OnCharHook(wxKeyEvent& evt)
       evt.Skip();
       return;
    }
-   
+
    if (IsOpenedFromEffectPanel())
       Close();
    else
@@ -978,19 +978,19 @@ wxBitmap EffectUIHost::CreateBitmap(const char * const xpm[], bool up, bool push
 {
    wxMemoryDC dc;
    wxBitmap pic(xpm);
-   
+
    wxBitmap mod(pic.GetWidth() + 6, pic.GetHeight() + 6, 24);
    dc.SelectObject(mod);
-   
+
 #if defined(__WXGTK__)
    wxColour newColour = wxSystemSettings::GetColour(wxSYS_COLOUR_BACKGROUND);
 #else
    wxColour newColour = wxSystemSettings::GetColour(wxSYS_COLOUR_BTNFACE);
 #endif
-   
+
    dc.SetBackground(wxBrush(newColour));
    dc.Clear();
-   
+
    int offset = 3;
    if (pusher)
    {
@@ -999,11 +999,11 @@ wxBitmap EffectUIHost::CreateBitmap(const char * const xpm[], bool up, bool push
          offset += 1;
       }
    }
-   
+
    dc.DrawBitmap(pic, offset, offset, true);
-   
+
    dc.SelectObject(wxNullBitmap);
-   
+
    return mod;
 }
 
@@ -1026,12 +1026,12 @@ void EffectUIHost::UpdateControls()
 void EffectUIHost::LoadUserPresets()
 {
    mUserPresets.clear();
-   
+
    GetConfigSubgroups(mEffectUIHost.GetDefinition(),
       PluginSettings::Private, UserPresetsGroup(wxEmptyString), mUserPresets);
-   
+
    std::sort( mUserPresets.begin(), mUserPresets.end() );
-   
+
    return;
 }
 
@@ -1041,7 +1041,7 @@ std::shared_ptr<EffectInstance> EffectUIHost::InitializeInstance()
    std::shared_ptr<EffectInstance> result;
 
    auto mpState = mwState.lock();
- 
+
    bool priorState = (mpState != nullptr);
    if (!priorState) {
       auto gAudioIO = AudioIO::Get();
@@ -1079,7 +1079,7 @@ std::shared_ptr<EffectInstance> EffectUIHost::InitializeInstance()
             }
          });
       }
-      
+
       mInitialized = true;
    }
    else
@@ -1139,209 +1139,6 @@ DialogFactoryResults EffectUI::DialogFactory(wxWindow &parent,
 #include "../SelectUtilities.h"
 #include "WaveTrack.h"
 #include "CommandManager.h"
-
-/// DoEffect() takes a PluginID and executes the associated effect.
-///
-/// At the moment flags are used only to indicate whether to prompt for
-//  parameters, whether to save the state to history and whether to allow
-/// 'Repeat Last Effect'.
-
-/* static */ bool EffectUI::DoEffect(
-   const PluginID & ID, const CommandContext &context, unsigned flags )
-{
-   AudacityProject &project = context.project;
-   auto &tracks = TrackList::Get( project );
-   auto &trackFactory = WaveTrackFactory::Get( project );
-   auto rate = ProjectRate::Get(project).GetRate();
-   auto &selectedRegion = ViewInfo::Get( project ).selectedRegion;
-   auto &commandManager = CommandManager::Get( project );
-   auto &viewport = Viewport::Get(project);
-   auto &window = GetProjectFrame(project);
-
-   const PluginDescriptor *plug = PluginManager::Get().GetPlugin(ID);
-
-   if (!plug || !PluginManager::IsPluginAvailable(*plug))
-   {
-      BasicUI::ShowMessageBox(
-         XO("This plugin could not be loaded.\nIt may have been deleted."),
-         BasicUI::MessageBoxOptions()
-            .Caption(XO("Plugin Error")));
-
-      return false;
-   }
-
-   EffectType type = plug->GetEffectType();
-
-   // Make sure there's no activity since the effect is about to be applied
-   // to the project's tracks.  Mainly for Apply during RTP, but also used
-   // for batch commands
-   if (flags & EffectManager::kConfigured)
-   {
-      ProjectAudioManager::Get( project ).Stop();
-      //Don't Select All if repeating Generator Effect
-      if (!(flags & EffectManager::kConfigured)) {
-         SelectUtilities::SelectAllIfNone(project);
-      }
-   }
-
-   auto nTracksOriginally = tracks.Size();
-   wxWindow *focus = wxWindow::FindFocus();
-   wxWindow *parent = nullptr;
-   if (focus != nullptr) {
-      parent = focus->GetParent();
-   }
-
-   bool success = false;
-   auto cleanup = finally( [&] {
-
-      if (!success) {
-         // For now, we're limiting realtime preview to a single effect, so
-         // make sure the menus reflect that fact that one may have just been
-         // opened.
-         CommandManager::Get(project).UpdateMenus( false );
-      }
-
-   } );
-
-   const auto range = tracks.Selected<const WaveTrack>();
-   bool anyTracks = !range.empty();
-   bool clean = std::all_of(range.begin(), range.end(),
-      [](const WaveTrack *t){ return t->GetEndTime() == 0; });
-
-   EffectManager & em = EffectManager::Get();
-
-   em.SetSkipStateFlag( false );
-   success = false;
-   if (auto effect = dynamic_cast<Effect*>(em.GetEffect(ID))) {
-      if (const auto pSettings = em.GetDefaultSettings(ID)) {
-         const auto pAccess =
-            std::make_shared<SimpleEffectSettingsAccess>(*pSettings);
-         const auto finder =
-         [effect, &window, pAccess, flags] (EffectSettings &settings)
-            -> std::optional<std::shared_ptr<EffectInstanceEx>>
-         {
-            // Prompting will be bypassed when applying an effect that has
-            // already been configured, e.g. repeating the last effect on a
-            // different selection.  Prompting may call EffectPreview
-            std::shared_ptr<EffectInstance> pInstance;
-            std::shared_ptr<EffectInstanceEx> pInstanceEx;
-            if ((flags & EffectManager::kConfigured) == 0 && pAccess) {
-               const auto pServices = dynamic_cast<EffectUIServices *>(effect);
-               if (!pServices ||
-                   !pServices->ShowHostInterface(*effect,
-                  window, DialogFactory, pInstance, *pAccess, true)
-               )
-                  return {};
-               else if (!(pInstanceEx =
-                  std::dynamic_pointer_cast<EffectInstanceEx>(pInstance)
-               ))
-                  return {};
-               else
-                  // Retrieve again after the dialog modified settings
-                  settings = pAccess->Get();
-            }
-            return { pInstanceEx };
-         };
-         pAccess->ModifySettings([&](EffectSettings &settings){
-            success = effect->DoEffect(settings, finder,
-               rate,
-               &tracks,
-               &trackFactory,
-               selectedRegion,
-               flags,
-               pAccess);
-            return nullptr;
-         });
-      }
-   }
-
-   if (!success)
-      return false;
-
-   if (em.GetSkipStateFlag())
-      flags = flags | EffectManager::kSkipState;
-
-   if (!(flags & EffectManager::kSkipState))
-   {
-      auto shortDesc = em.GetCommandName(ID);
-      auto longDesc = em.GetCommandDescription(ID);
-      ProjectHistory::Get( project ).PushState(longDesc, shortDesc);
-   }
-
-   if (!(flags & EffectManager::kDontRepeatLast))
-   {
-      // Remember a successful generator, effect, analyzer, or tool Process
-         auto shortDesc = em.GetCommandName(ID);
-         /* i18n-hint: %s will be the name of the effect which will be
-          * repeated if this menu item is chosen */
-         auto lastEffectDesc = XO("Repeat %s").Format(shortDesc);
-         switch ( type ) {
-         case EffectTypeGenerate:
-            commandManager.Modify(wxT("RepeatLastGenerator"), lastEffectDesc);
-            commandManager.mLastGenerator = ID;
-            commandManager.mRepeatGeneratorFlags = EffectManager::kConfigured;
-            break;
-         case EffectTypeProcess:
-            commandManager.Modify(wxT("RepeatLastEffect"), lastEffectDesc);
-            commandManager.mLastEffect = ID;
-            commandManager.mRepeatEffectFlags = EffectManager::kConfigured;
-            break;
-         case EffectTypeAnalyze:
-            commandManager.Modify(wxT("RepeatLastAnalyzer"), lastEffectDesc);
-            commandManager.mLastAnalyzer = ID;
-            commandManager.mLastAnalyzerRegistration =
-               CommandManager::repeattypeplugin;
-            commandManager.mRepeatAnalyzerFlags = EffectManager::kConfigured;
-            break;
-         case EffectTypeTool:
-            commandManager.Modify(wxT("RepeatLastTool"), lastEffectDesc);
-            commandManager.mLastTool = ID;
-            commandManager.mLastToolRegistration =
-               CommandManager::repeattypeplugin;
-            commandManager.mRepeatToolFlags = EffectManager::kConfigured;
-            if (shortDesc == NYQUIST_PROMPT_NAME) {
-               commandManager.mRepeatToolFlags = EffectManager::kRepeatNyquistPrompt;  //Nyquist Prompt is not configured
-            }
-            break;
-      }
-   }
-
-   //STM:
-   //The following automatically re-zooms after sound was generated.
-   // IMO, it was disorienting, removing to try out without re-fitting
-   //mchinen:12/14/08 reapplying for generate effects
-   if (type == EffectTypeGenerate)
-   {
-      if (!anyTracks || (clean && selectedRegion.t0() == 0.0))
-         viewport.ZoomFitHorizontally();
-   }
-
-   // PRL:  Redraw explicitly because sometimes history push is skipped
-   viewport.Redraw();
-
-   if (focus != nullptr && focus->GetParent()==parent) {
-      focus->SetFocus();
-   }
-
-   // A fix for Bug 63
-   // New tracks added?  Scroll them into view so that user sees them.
-   // Don't care what track type.  An analyser might just have added a
-   // Label track and we want to see it.
-   if (tracks.Size() > nTracksOriginally) {
-      viewport.ScrollToBottom();
-   }
-   else {
-      auto pTrack = *tracks.Selected().begin();
-      if (!pTrack)
-         pTrack = *tracks.begin();
-      if (pTrack) {
-         TrackFocus::Get(project).Set(pTrack);
-         Viewport::Get(project).ShowTrack(*pTrack);
-      }
-   }
-
-   return true;
-}
 
 ///////////////////////////////////////////////////////////////////////////////
 BEGIN_EVENT_TABLE(EffectDialog, wxDialogWrapper)
