+++ src/effects/nyquist/Nyquist.cpp
@@ -53,7 +53,7 @@ effects from this one class.
 
 #include "../../LabelTrack.h"
 #include "NoteTrack.h"
-#include "../../ShuttleGetDefinition.h"
+#include "ShuttleGetDefinition.h"
 #include "../../prefs/GUIPrefs.h"
 #include "../../prefs/SpectrogramSettings.h"
 #include "../../tracks/playabletrack/wavetrack/ui/WaveChannelView.h"
@@ -61,9 +61,9 @@ effects from this one class.
 #include "../../widgets/NumericTextCtrl.h"
 #include "../../widgets/valnum.h"
 #include "../EffectEditor.h"
-#include "../EffectManager.h"
 #include "AudacityMessageBox.h"
 #include "BasicUI.h"
+#include "EffectManager.h"
 #include "FileNames.h"
 #include "Languages.h"
 #include "PluginManager.h"
